+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Inline `@import` rules in `tailwindcss/index.css` at publish time for better performance ([#13233](https://github.com/tailwindlabs/tailwindcss/pull/13233))
 
+### Fixed
+
+- Validate bare values ([#13245](https://github.com/tailwindlabs/tailwindcss/pull/13245))
+
 ## [4.0.0-alpha.9] - 2024-03-13
 
 ### Added